% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_list_hi_ts_for_qualification_type}
\alias{mturk_list_hi_ts_for_qualification_type}
\title{The ListHITsForQualificationType operation returns the HITs that use the
given Qualification type for a Qualification requirement}
\usage{
mturk_list_hi_ts_for_qualification_type(QualificationTypeId, NextToken,
  MaxResults)
}
\arguments{
\item{QualificationTypeId}{[required] The ID of the Qualification type to use when querying HITs.}

\item{NextToken}{Pagination Token}

\item{MaxResults}{Limit the number of results returned.}
}
\description{
The \code{ListHITsForQualificationType} operation returns the HITs that use
the given Qualification type for a Qualification requirement. The
operation returns HITs of any status, except for HITs that have been
deleted with the \code{DeleteHIT} operation or that have been auto-deleted.
}
\section{Request syntax}{
\preformatted{mturk$list_hi_ts_for_qualification_type(
  QualificationTypeId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
